Grpc service added and Book service

+++ schoolAPI/Models/StudentStudyProgramme.cs
@@ -5,10 +5,10 @@ namespace schoolAPI.Models
     public class StudentStudyProgramme
     {
         [ForeignKey("StudentId")]
-        public long StudentId { get; set; }
+        public int StudentId { get; set; }
         public Student Student { get; set; }
         [ForeignKey("StudyProgrammeId")]
-        public long StudyProgrammeId { get; set; }
+        public int StudyProgrammeId { get; set; }
         public StudyProgramme StudyProgramme { get; set; }
 
     }
